+++ CHANGELOG.md
@@ -33,6 +33,9 @@
   * New functions: `pyjlarray`, `pyjldict`, `pyjlset`.
 
 ## Unreleased
+* Bug fix: avoid precompilation cache miss when `check_bounds` option set.
+
+## 0.9.35 (2026-06-08)
 * Add option `lib` to JuliaCall. Setting this will skip the discovery subprocess.
 * Add support for using a system image in `juliacall` that has `PythonCall` baked in.
 * Add method `pynext(x, d)` to return a default value `d` if there are no more elements.